Refactor logger location
