+++ lib/api/github.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const ghGot = require('gh-got');
 
 const config = {};
