+++ lib/api/gitlab.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const glGot = require('gl-got');
 
 const config = {};
