+++ lib/api/npm.js
@@ -4,7 +4,7 @@ const got = require('got');
 const url = require('url');
 const registryUrl = require('registry-url');
 const registryAuthToken = require('registry-auth-token');
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 
 module.exports = {
   setNpmrc,
