+++ lib/config/file.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const path = require('path');
 
 module.exports = {
