+++ lib/config/index.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const stringify = require('json-stringify-pretty-compact');
 const githubApi = require('../api/github');
 const gitlabApi = require('../api/gitlab');
