+++ lib/helpers/changelog.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const changelog = require('changelog');
 
 module.exports = {
