+++ lib/helpers/github-app.js
@@ -1,5 +1,5 @@
 const jwt = require('jsonwebtoken');
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const ghApi = require('../api/github');
 
 module.exports = {
