+++ lib/helpers/logger.js
@@ -1,5 +1,5 @@
 const bunyan = require('bunyan');
-const CliHelper = require('./helpers/cli');
+const CliHelper = require('./cli');
 
 const cliHelper = new CliHelper();
 cliHelper.pipe(process.stdout);
