+++ lib/helpers/npm.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const fs = require('fs');
 const cp = require('child_process');
 const tmp = require('tmp');
