+++ lib/helpers/package-json.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const _ = require('lodash');
 
 module.exports = {
