+++ lib/helpers/versions.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const semver = require('semver');
 const stable = require('semver-stable');
 const _ = require('lodash');
