+++ lib/index.js
@@ -1,4 +1,4 @@
-const logger = require('./logger');
+const logger = require('./helpers/logger');
 const configParser = require('./config');
 const repositoryWorker = require('./workers/repository');
