+++ lib/workers/branch.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const stringifyConfig = require('../helpers/stringify-config');
 const handlebars = require('handlebars');
 const packageJsonHelper = require('../helpers/package-json');
