+++ lib/workers/packages.js
@@ -4,7 +4,7 @@ const handlebars = require('handlebars');
 const prWorker = require('./pr');
 const branchWorker = require('./branch');
 // Logger
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 // API
 const npmApi = require('../api/npm');
 // Helperss
