+++ lib/workers/pr.js
@@ -1,4 +1,4 @@
-const logger = require('../logger');
+const logger = require('../helpers/logger');
 const handlebars = require('handlebars');
 const changelogHelper = require('../helpers/changelog');
 const stringifyConfig = require('../helpers/stringify-config');
